[skip ci] feat, wip: ':parse' command for the repl
